pyproject.toml
[project]
name = "git-merge-conflict-env"
version = "1.0.0"
description = "RL environment for training AI agents to resolve git merge conflicts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.3",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
server = "git_merge_conflict_env.server.app:main"

[tool.setuptools]
packages = ["git_merge_conflict_env", "git_merge_conflict_env.server"]
package-dir = { "git_merge_conflict_env" = "git_merge_conflict_env", "git_merge_conflict_env.server" = "git_merge_conflict_env/server" }
